% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_est.R
\name{coef_est}
\alias{coef_est}
\title{Individual level effect estimation}
\usage{
coef_est(id, x, y, weights, nfolds = 5, nvisit = 6)
}
\arguments{
\item{id}{The vector of individual level ID's from the original dataset.}

\item{x}{The covariate matrix.}

\item{y}{The outcome variable.}

\item{weights}{Observation weights.}

\item{nfolds}{The number of folds.}

\item{nvisit}{The total number of time points.}
}
\value{
The estimated coefficients for subject level covariates.
}
\description{
This function is designed to estimate subject-level coefficients
}
