% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLNet.R
\name{JLNet}
\alias{JLNet}
\title{JLNet}
\usage{
JLNet(
  x,
  y,
  subj.id,
  hosp.id,
  visit,
  data,
  models,
  family,
  nfolds = 5,
  nvisit,
  k = c(2:30),
  nstart = 25
)
}
\arguments{
\item{x}{The covariate matrix.}

\item{y}{The observed outcome variable.}

\item{subj.id}{The subject ID.}

\item{hosp.id}{The hospital ID.}

\item{visit}{The longitudinal information}

\item{data}{The original dataset}

\item{models}{A formula object describing the missing mechanism.}

\item{family}{The distribution family that the missing indicator belongs to.}

\item{nfolds}{The number of folds.}

\item{nvisit}{The total number of time points.}

\item{k}{The number of clusters.}

\item{nstart}{The number of random sets chosen for clustering.}
}
\value{
A list of 1. the estimated individual-level effect, 2. The estimated coefficients for subject level covariates.
}
\description{
This function is designed to estimate subject-level coefficients
}
