% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss.R
\name{miss}
\alias{miss}
\title{Drop-out missing model}
\usage{
miss(models, data, visit, subj.id, family)
}
\arguments{
\item{models}{A formula object describing the missing mechanism.}

\item{data}{A data frame containing the variables in the model.}

\item{visit}{The variable of longitudinal information.}

\item{family}{The distribution family that the missing indicator belongs to.}

\item{id}{The individual level ID.}
}
\value{
A list with two items: The estimated coefficients of the missing model and inverse probability weights
}
\description{
Generate inverse probability weight (IPW) for longitudinal data with drop-out situation
}
