% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gmm.R
\name{new_gmm}
\alias{new_gmm}
\title{new_gmm}
\usage{
new_gmm(k, thetahat, sigma_all, nstart = 25)
}
\arguments{
\item{k}{The number of clusters.}

\item{thetahat}{The initial estimate of facility-level effect.}

\item{sigma_all}{The standard error of thetahat.}

\item{nstart}{The number of random sets chosen for clustering.}
}
\value{
A list of two items: 1. a vector indicating the cluster allocated to and 2. a vector of fitted facility-level effect.
}
\description{
This function is designed to estimate the facility level effects.
}
