% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gmm_bic.R
\name{new_gmm_bic}
\alias{new_gmm_bic}
\title{new_gmm_bic}
\usage{
new_gmm_bic(k, thetahat, sigma_all, nstart = 25)
}
\arguments{
\item{k}{The number of clusters.}

\item{thetahat}{The initial estimate of facility-level effect.}

\item{sigma_all}{The standard error of thetahat.}

\item{nstart}{The number of random sets chosen for clustering.}
}
\value{
Bayesian information criterion (BIC) of the fitted model.
}
\description{
This function is supposed to pinpoint the model minimizing the BIC
}
