% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_est.R
\name{theta_est}
\alias{theta_est}
\title{Hospital-level effect initial estimation.}
\usage{
theta_est(x, y, subj.id, hosp.id, data, beta, weights)
}
\arguments{
\item{x}{The covariate matrix.}

\item{y}{The outcome variable.}

\item{subj.id}{The subject level ID.}

\item{hosp.id}{The hospital level ID.}

\item{data}{The original dataset.}

\item{beta}{The estimation of individual-level effect}

\item{weights}{The weight used to adjust for drop-out}
}
\value{
A list of two items: the initial estimate of hospital-level effect and the standard error of the estimation.
}
\description{
This function is designed to obtain the initial estimate of hospital-level effect.
}
