% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\title{transform function}
\usage{
trans(x.data, y.data, weight, hosp.id)
}
\arguments{
\item{x.data}{The original predictor matrix}

\item{y.data}{The original response variable}

\item{weight}{weights}

\item{hosp.id}{The hospital ID.}
}
\value{
A list with two items: The transformed predictor matrix and the transformed response variable
}
\description{
Transform predictor matrix and response variable
}
